Initialized the project structure for the Metarigger.
A new visual studio python project was added in /source
and was configured to have access to maya libraries and to use mayaapi as its python interpreter.

The project is structured around the Metarigger package. A second package was already added under Metarigger as Metanodes.

Furthermore, a test directory, that mimics the project structure, was added to contain all the tests for the it.