The basic core of the Metanode class were built along with its test cases.
A Metanode class is a class that wraps a maya node to store data about it and its construction.
It contains informations about about how the wrapped node relates to other nodes to simplify graph traversal  and rig building.

A metanode is represented in the maya graph with an actual uknown node instance that stores the needed informations in its attribute.