The core interface for Metanode Connections was added.
Connections are one of the data types that a Metanode can store on its node representation.
Connections are used as a way to represent relationships with other nodes.
A connection can be added with the addConnection method.
For each connection of the metanode an equivalent walkTo{ConnectionName} is added to the Metanode class instance that returns the node that is related to the connection.