The core interface for metanode options was added.
An option is on of the data types the metanode use to store its informations.
It represents options with on/off states that can be used for branching during build time.
Every time an option is added two methods are added to the Metanode class instance, a getter and a setter.
For an option named {name} the getter is added under the name is{name} and the setter under the name set{name}

The MetanodeConnections.py test file was renamed to testMetanodeConnecions.py to respect the currently used naming convention.