Added a bit of expressiveness to the test.
A new package was added to the Metarigger : Utils.
This is where every generic module that has some use will go.
In it a TestUtils package with a MetariggerTestCase module was added.
The MetariggerTestCase contatins an omonymn class that serves the purpose of streamlining some of the common assert that are appearing in tests and giving a better expression of their purpose.

The tests were update to use the new class.