The core counter interface for the metanode class was implemented.
A counter is a float attribute that is used to store variable amounts informations.
It can, for example, be used to store the number of joints in a spine module.

When a counter is added two methods, a getter and a setter, are added to the Metanode Class Instance.
For a counter with name as {name} the getter is called howMany{name} and the setter is called setHowMany{name}.
A counter always has a min and max range. If a value outside this range is tried to be setted on the counter it is clamped to the nearest valid amount.