The core interface for Information attribute was added.
AN information store generic values as strings.
Every time an information is added with the addInformation methods a string attribute is added to the metanode maya instance while two methods are added to the Metanode class instance.
For an Information named {name} a getter method, which{name}, and a setter method, setWhich{name}, are added.

The test file for counters was renamed to follow the correct naming convention and some typos were corrected from it.
Furthermore, the metanode in file documentation was updated to the current specifications.