The core Widget class was implemented.
A widget is a metanode that is represented in the graph by a dag node instead of a dg node.
As such it can be visually repositioned in the scene.
This can be used to position skeletons or other important pieces to before the build process to adjust them to the rigged mesh.

The core functioning of the Widget class is currently pretty bare and is expected to be expaded upon in the near future.

Some changes were made to the Metanode class to support the new Widgets.
The Metanode maya instance is now built trough the _buildMetanode method. This was done to permit the customization of the nodetype of the maya representation of the class for Metanode derived classes.

The _instance member was removed from the Metanode class now exists only as a default connection.
As such, the default build method was changed to connect the messa attribute of the newly created node and the _instance connection on the metanode representation.
Furthermore, the addConnection method now adds an additional method that checks if a Connection has any connection currently going on in the graph.